removes unncessary AppFonts
